Cleaning up Jablko in preparation for work on Jarmuz-RGB-Light
