+++ builtin/demo/makefile
@@ -1,0 +1,9 @@
+run: demomod
+	./demomod
+
+demomod: demomod.go jablkohelpers.go
+	echo "ASDASDASDASDASDASDASDASDASDS"
+	go build -o demomod .
+
+build:
+	go build -o demomod .
