--- builtin/test/makefile
@@ -1,9 +1,0 @@
-run: jmod_test
-	./jmod_test
-
-jmod_test: testmod.go jablkohelpers.go
-	echo "ASDASDASDASDASDASDASDASDASDS"
-	go build -o jmod_test .
-
-build:
-	go build -o jmod_test .
