+++ jmods.json
@@ -8,7 +8,7 @@
             }
         }
     },
-    "builtin/test": {
+    "builtin/demo": {
         "portUDP": 41111,
         "instances": {
             "inst0": {
@@ -16,15 +16,6 @@
             },
             "inst99": {
                 "value": 49
-            }
-        }
-    },
-    "github.com/ccoverstreet/Jarmuz-RGB-Light@latest": {
-        "instances": {
-            "inst1": {
-                "lightIPs": [
-                    "10.0.0.10"
-                ]
             }
         }
     }
